+++ docs/html/search/all_3.js
@@ -1,8 +1,8 @@
 var searchData=
 [
-  ['digitalwrite_7',['digitalWrite',['../classace__segment_1_1GpioInterface.html#ae87dcb5fa1c1245fc5747abfeebac7b4',1,'ace_segment::GpioInterface']]],
-  ['directfast4module_8',['DirectFast4Module',['../classace__segment_1_1DirectFast4Module.html',1,'ace_segment']]],
-  ['directmodule_9',['DirectModule',['../classace__segment_1_1DirectModule.html',1,'ace_segment']]],
-  ['disablegroup_10',['disableGroup',['../classace__segment_1_1LedMatrixBase.html#a392418ec3a4e763920e21acc763e8848',1,'ace_segment::LedMatrixBase::disableGroup()'],['../classace__segment_1_1LedMatrixDualHc595.html#a0a08cfdd1de0cbdb20b9e66b22627c86',1,'ace_segment::LedMatrixDualHc595::disableGroup()']]],
-  ['draw_11',['draw',['../classace__segment_1_1LedMatrixBase.html#ad9d2bc50b50ab4b2c16c1ef53b3397d2',1,'ace_segment::LedMatrixBase::draw()'],['../classace__segment_1_1LedMatrixDualHc595.html#ac191206242ca55ebb31652de7993277e',1,'ace_segment::LedMatrixDualHc595::draw()']]]
+  ['digitalwrite_10',['digitalWrite',['../classace__segment_1_1GpioInterface.html#ae87dcb5fa1c1245fc5747abfeebac7b4',1,'ace_segment::GpioInterface']]],
+  ['directfast4module_11',['DirectFast4Module',['../classace__segment_1_1DirectFast4Module.html',1,'ace_segment']]],
+  ['directmodule_12',['DirectModule',['../classace__segment_1_1DirectModule.html',1,'ace_segment']]],
+  ['disablegroup_13',['disableGroup',['../classace__segment_1_1LedMatrixBase.html#a392418ec3a4e763920e21acc763e8848',1,'ace_segment::LedMatrixBase::disableGroup()'],['../classace__segment_1_1LedMatrixDualHc595.html#a0a08cfdd1de0cbdb20b9e66b22627c86',1,'ace_segment::LedMatrixDualHc595::disableGroup()']]],
+  ['draw_14',['draw',['../classace__segment_1_1LedMatrixBase.html#ad9d2bc50b50ab4b2c16c1ef53b3397d2',1,'ace_segment::LedMatrixBase::draw()'],['../classace__segment_1_1LedMatrixDualHc595.html#ac191206242ca55ebb31652de7993277e',1,'ace_segment::LedMatrixDualHc595::draw()']]]
 ];