+++ docs/html/search/all_5.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['flush_15',['flush',['../classace__segment_1_1Ht16k33Module.html#ad9225465a7ce3fa621eb8505edb77f6a',1,'ace_segment::Ht16k33Module::flush()'],['../classace__segment_1_1Max7219Module.html#af8b588c48e73e6883948eb8159982a1c',1,'ace_segment::Max7219Module::flush()'],['../classace__segment_1_1Tm1637Module.html#af3a100b0d07125af76aaf8ea10bc597c',1,'ace_segment::Tm1637Module::flush()']]],
-  ['flushincremental_16',['flushIncremental',['../classace__segment_1_1Tm1637Module.html#a1bf9dcdafb5b1a08a18da697d29e5235',1,'ace_segment::Tm1637Module']]]
+  ['flush_18',['flush',['../classace__segment_1_1Ht16k33Module.html#ad9225465a7ce3fa621eb8505edb77f6a',1,'ace_segment::Ht16k33Module::flush()'],['../classace__segment_1_1Max7219Module.html#af8b588c48e73e6883948eb8159982a1c',1,'ace_segment::Max7219Module::flush()'],['../classace__segment_1_1Tm1637Module.html#af3a100b0d07125af76aaf8ea10bc597c',1,'ace_segment::Tm1637Module::flush()']]],
+  ['flushincremental_19',['flushIncremental',['../classace__segment_1_1Tm1637Module.html#a1bf9dcdafb5b1a08a18da697d29e5235',1,'ace_segment::Tm1637Module']]]
 ];