+++ docs/html/search/all_b.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['max7219module_49',['Max7219Module',['../classace__segment_1_1Max7219Module.html',1,'ace_segment::Max7219Module&lt; T_SPII, T_DIGITS &gt;'],['../classace__segment_1_1Max7219Module.html#a3ba94c7d5ba3583dfc22999c22d86f12',1,'ace_segment::Max7219Module::Max7219Module()']]],
-  ['micros_50',['micros',['../classace__segment_1_1ClockInterface.html#afe559a873d3dae65e6430162702a5ee6',1,'ace_segment::ClockInterface']]],
-  ['millis_51',['millis',['../classace__segment_1_1ClockInterface.html#a53a5faea52b0eaeca0bfcb78c554a8d4',1,'ace_segment::ClockInterface']]]
+  ['max7219module_51',['Max7219Module',['../classace__segment_1_1Max7219Module.html',1,'ace_segment::Max7219Module&lt; T_SPII, T_DIGITS &gt;'],['../classace__segment_1_1Max7219Module.html#a3ba94c7d5ba3583dfc22999c22d86f12',1,'ace_segment::Max7219Module::Max7219Module()']]],
+  ['micros_52',['micros',['../classace__segment_1_1ClockInterface.html#afe559a873d3dae65e6430162702a5ee6',1,'ace_segment::ClockInterface']]],
+  ['millis_53',['millis',['../classace__segment_1_1ClockInterface.html#a53a5faea52b0eaeca0bfcb78c554a8d4',1,'ace_segment::ClockInterface']]]
 ];