+++ docs/html/search/classes_3.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['hc595module_102',['Hc595Module',['../classace__segment_1_1Hc595Module.html',1,'ace_segment']]],
-  ['ht16k33module_103',['Ht16k33Module',['../classace__segment_1_1Ht16k33Module.html',1,'ace_segment']]],
-  ['hybridmodule_104',['HybridModule',['../classace__segment_1_1HybridModule.html',1,'ace_segment']]]
+  ['hc595module_106',['Hc595Module',['../classace__segment_1_1Hc595Module.html',1,'ace_segment']]],
+  ['ht16k33module_107',['Ht16k33Module',['../classace__segment_1_1Ht16k33Module.html',1,'ace_segment']]],
+  ['hybridmodule_108',['HybridModule',['../classace__segment_1_1HybridModule.html',1,'ace_segment']]]
 ];