+++ docs/html/search/classes_4.js
@@ -1,11 +1,11 @@
 var searchData=
 [
-  ['ledmatrixbase_105',['LedMatrixBase',['../classace__segment_1_1LedMatrixBase.html',1,'ace_segment']]],
-  ['ledmatrixdirect_106',['LedMatrixDirect',['../classace__segment_1_1LedMatrixDirect.html',1,'ace_segment']]],
-  ['ledmatrixdirect_3c_20gpiointerface_20_3e_107',['LedMatrixDirect&lt; GpioInterface &gt;',['../classace__segment_1_1LedMatrixDirect.html',1,'ace_segment']]],
-  ['ledmatrixdualhc595_108',['LedMatrixDualHc595',['../classace__segment_1_1LedMatrixDualHc595.html',1,'ace_segment']]],
-  ['ledmatrixsinglehc595_109',['LedMatrixSingleHc595',['../classace__segment_1_1LedMatrixSingleHc595.html',1,'ace_segment']]],
-  ['ledmatrixsinglehc595_3c_20t_5fspii_2c_20gpiointerface_20_3e_110',['LedMatrixSingleHc595&lt; T_SPII, GpioInterface &gt;',['../classace__segment_1_1LedMatrixSingleHc595.html',1,'ace_segment']]],
-  ['ledmodule_111',['LedModule',['../classace__segment_1_1LedModule.html',1,'ace_segment']]],
-  ['levelwriter_112',['LevelWriter',['../classace__segment_1_1LevelWriter.html',1,'ace_segment']]]
+  ['ledmatrixbase_109',['LedMatrixBase',['../classace__segment_1_1LedMatrixBase.html',1,'ace_segment']]],
+  ['ledmatrixdirect_110',['LedMatrixDirect',['../classace__segment_1_1LedMatrixDirect.html',1,'ace_segment']]],
+  ['ledmatrixdirect_3c_20gpiointerface_20_3e_111',['LedMatrixDirect&lt; GpioInterface &gt;',['../classace__segment_1_1LedMatrixDirect.html',1,'ace_segment']]],
+  ['ledmatrixdualhc595_112',['LedMatrixDualHc595',['../classace__segment_1_1LedMatrixDualHc595.html',1,'ace_segment']]],
+  ['ledmatrixsinglehc595_113',['LedMatrixSingleHc595',['../classace__segment_1_1LedMatrixSingleHc595.html',1,'ace_segment']]],
+  ['ledmatrixsinglehc595_3c_20t_5fspii_2c_20gpiointerface_20_3e_114',['LedMatrixSingleHc595&lt; T_SPII, GpioInterface &gt;',['../classace__segment_1_1LedMatrixSingleHc595.html',1,'ace_segment']]],
+  ['ledmodule_115',['LedModule',['../classace__segment_1_1LedModule.html',1,'ace_segment']]],
+  ['levelwriter_116',['LevelWriter',['../classace__segment_1_1LevelWriter.html',1,'ace_segment']]]
 ];