+++ docs/html/search/classes_5.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['max7219module_113',['Max7219Module',['../classace__segment_1_1Max7219Module.html',1,'ace_segment']]]
+  ['max7219module_117',['Max7219Module',['../classace__segment_1_1Max7219Module.html',1,'ace_segment']]]
 ];