+++ docs/html/search/classes_6.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['numberwriter_114',['NumberWriter',['../classace__segment_1_1NumberWriter.html',1,'ace_segment']]]
+  ['numberwriter_118',['NumberWriter',['../classace__segment_1_1NumberWriter.html',1,'ace_segment']]]
 ];