+++ docs/html/search/classes_7.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['patternwriter_115',['PatternWriter',['../classace__segment_1_1PatternWriter.html',1,'ace_segment']]]
+  ['patternwriter_119',['PatternWriter',['../classace__segment_1_1PatternWriter.html',1,'ace_segment']]]
 ];