+++ docs/html/search/classes_9.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['temperaturewriter_123',['TemperatureWriter',['../classace__segment_1_1TemperatureWriter.html',1,'ace_segment']]],
-  ['tm1637module_124',['Tm1637Module',['../classace__segment_1_1Tm1637Module.html',1,'ace_segment']]]
+  ['temperaturewriter_127',['TemperatureWriter',['../classace__segment_1_1TemperatureWriter.html',1,'ace_segment']]],
+  ['tm1637module_128',['Tm1637Module',['../classace__segment_1_1Tm1637Module.html',1,'ace_segment']]]
 ];