+++ docs/html/search/searchdata.js
@@ -4,8 +4,7 @@ var indexSectionsWithContent =
   1: "cdghlmnpst",
   2: "bcdefghilmnprstw",
   3: "k",
-  4: "h",
-  5: "a"
+  4: "a"
 };
 
 var indexSectionNames =
@@ -14,8 +13,7 @@ var indexSectionNames =
   1: "classes",
   2: "functions",
   3: "variables",
-  4: "typedefs",
-  5: "pages"
+  4: "pages"
 };
 
 var indexSectionLabels =
@@ -24,7 +22,6 @@ var indexSectionLabels =
   1: "Classes",
   2: "Functions",
   3: "Variables",
-  4: "Typedefs",
-  5: "Pages"
+  4: "Pages"
 };