+++ docs/html/search/typedefs_0.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['hexchar_5ft_208',['hexchar_t',['../classace__segment_1_1NumberWriter.html#abf1fc3ae5c324d1de2e3d5f196e5a0a8',1,'ace_segment::NumberWriter']]]
-];