+++ docs/html/search/variables_0.js
@@ -1,11 +1,6 @@
 var searchData=
 [
-  ['kcharminus_200',['kCharMinus',['../classace__segment_1_1ClockWriter.html#ab535943b0a616a41bf686b74f9a9fc31',1,'ace_segment::ClockWriter::kCharMinus()'],['../classace__segment_1_1NumberWriter.html#a039ac6a2528399ac73b384f6227d3e6a',1,'ace_segment::NumberWriter::kCharMinus()']]],
-  ['kcharspace_201',['kCharSpace',['../classace__segment_1_1ClockWriter.html#a5432977b4275f45adcc3764296a7c8ce',1,'ace_segment::ClockWriter::kCharSpace()'],['../classace__segment_1_1NumberWriter.html#ad5907a93c3f37b3cef81d3427688d18c',1,'ace_segment::NumberWriter::kCharSpace()']]],
-  ['knumhexchars_202',['kNumHexChars',['../classace__segment_1_1NumberWriter.html#a2da066eed77d440c8d2df88ad38407bb',1,'ace_segment::NumberWriter']]],
-  ['kpatterna_203',['kPatternA',['../classace__segment_1_1ClockWriter.html#a53759004793ba6bcc90408725829434f',1,'ace_segment::ClockWriter']]],
-  ['kpatternc_204',['kPatternC',['../classace__segment_1_1TemperatureWriter.html#ab708932843ab2f48fe22f1a95a1ea156',1,'ace_segment::TemperatureWriter']]],
-  ['kpatterndegree_205',['kPatternDegree',['../classace__segment_1_1TemperatureWriter.html#a1be9f982f9466b3afa0d01051dc04855',1,'ace_segment::TemperatureWriter']]],
-  ['kpatternf_206',['kPatternF',['../classace__segment_1_1TemperatureWriter.html#aab41ee81d8fe308c9d0c0d9c1e3c381d',1,'ace_segment::TemperatureWriter']]],
-  ['kpatternp_207',['kPatternP',['../classace__segment_1_1ClockWriter.html#aba763a1a69889af49dc7a34c972a309b',1,'ace_segment::ClockWriter']]]
+  ['kpatternc_214',['kPatternC',['../classace__segment_1_1TemperatureWriter.html#ad4c0e6ee19322219691196f9477c6d77',1,'ace_segment::TemperatureWriter']]],
+  ['kpatterndegree_215',['kPatternDegree',['../classace__segment_1_1TemperatureWriter.html#a5dc20ceed565269e7cf2b87abf3a7b5b',1,'ace_segment::TemperatureWriter']]],
+  ['kpatternf_216',['kPatternF',['../classace__segment_1_1TemperatureWriter.html#aa534955f5723086e542dfdb334412e3d',1,'ace_segment::TemperatureWriter']]]
 ];