+++ examples/AutoBenchmark/generate_readme.py
@@ -55,7 +55,7 @@ LED module:
 * `Ht16k33Module::flush()`
     * Sends out the buffered digits using I2C
 
-**Version**: AceSegment v0.8.1
+**Version**: AceSegment v0.8.2
 
 **DO NOT EDIT**: This file was auto-generated using `make README.md`.
 
@@ -209,6 +209,13 @@ number of `TimingStats::update()` calls that were made.
       `<Wire.h>`, until we get to very fast 32-bit processors like the ESP32 and
       Teensy 3.2, where `SimpleWireInterface` becomes competitive with 400 kHz
       of `<Wire.h>`.
+
+**v0.8+**
+
+* Remove `virtual` keyword from `LedModule` methods.
+    * No significant changes in execution time.
+* Templatize Writer classes on `T_LED_MODULE` instead of hardcoded `LedModule`.
+    * No significant changes in execution time.
 
 ## Results