+++ examples/MemoryBenchmark/generate_readme.py
@@ -40,7 +40,7 @@ by the runtime environment of the processor. For example, it often seems like
 the ESP8266 allocates flash memory in blocks of a certain quantity, so the
 calculated flash size can jump around in unexpected ways.
 
-**Version**: AceSegment v0.8.1
+**Version**: AceSegment v0.8.2
 
 **DO NOT EDIT**: This file was auto-generated using `make README.md`.
 
@@ -219,6 +219,22 @@ before substantional refactoring in 2021.
       bytes, but decreases by a few bytes in a few cases.
     * The 32-bit processors have so much more flash memory than 8-bit
       processors, I think this tradeoff is worth it.
+
+**v0.8.2**
+
+* Remove `virtual` keyword from `LedModule` methods.
+    * Decreases flash usage by 60 bytes for `Tm1637Module`, 14 bytes for
+      `Max7219Module`, 32 bytes for `Ht16k33Module`, and 2-14 bytes for
+      `Hc595Module`.
+    * Decreases static ram usage by 7-8 bytes for all Module classes.
+    * Further decreases flash usage by 10-70 bytes for various Writer classes.
+* Templatize Writer classes on `T_LED_MODULE` instead of hardcoding it to
+  `LedModule`.
+    * Seems to reduce flash size of some Writer classes on some platforms by
+      hundreds of bytes, I think because methods can be better inlined, and
+      unused methods are not compiled and linked in.
+* Add `isFlushRequired()` and clear appropriate flags after `flush()`.
+    * Increases flash consumption by about 8 bytes on AVR.
 
 ## Results