+++ src/ace_segment/scanning/ScanningModule.h
@@ -30,7 +30,8 @@ SOFTWARE.
 #include "../hw/ClockInterface.h" // ClockInterface
 #include "../LedModule.h"
 
-class ScanningModuleTest_displayCurrentField;
+class ScanningModuleTest_isAnyDigitDirty;
+class ScanningModuleTest_isBrightnessDirty;
 
 namespace ace_segment {
 
@@ -94,7 +95,7 @@ class ScanningModule : public LedModule {
         const T_LM& ledMatrix,
         uint8_t framesPerSecond
     ):
-        LedModule(T_DIGITS),
+        LedModule(mPatterns, T_DIGITS),
         mLedMatrix(ledMatrix),
         mFramesPerSecond(framesPerSecond)
     {}
@@ -106,6 +107,8 @@ class ScanningModule : public LedModule {
      * time.
      */
     void begin() {
+      LedModule::begin();
+
       // Set up durations for the renderFieldWhenReady() polling function.
       mMicrosPerField = (uint32_t) 1000000UL / getFieldsPerSecond();
       mLastRenderFieldMicros = T_CI::micros();
@@ -117,6 +120,7 @@ class ScanningModule : public LedModule {
       mPattern = 0;
 
       // Set initial patterns and global brightness.
+      mIsDigitBrightnessDirty = false;
       mLedMatrix.clear();
       if (T_SUBFIELDS > 1) {
         setBrightness(T_SUBFIELDS / 2); // half brightness
@@ -125,33 +129,8 @@ class ScanningModule : public LedModule {
 
 
     /** A no-op end() function for consistency with other classes. */
-    void end() {}
-
-    //-----------------------------------------------------------------------
-    // Implement the LedModule interface.
-    //-----------------------------------------------------------------------
-
-    /** Get the number of digits. */
-    uint8_t getNumDigits() const { return T_DIGITS; }
-
-    void setPatternAt(uint8_t pos, uint8_t pattern) override {
-      mPatterns[pos] = pattern;
-    }
-
-    uint8_t getPatternAt(uint8_t pos) override {
-      return mPatterns[pos];
-    }
-
-    /**
-     * @copydoc
-     *
-     * See the documentation for setBrightnessAt() for information about the
-     * range of values of `brightness` and how it is interpreted.
-     */
-    void setBrightness(uint8_t brightness) override {
-      for (uint8_t i = 0; i < T_DIGITS; i++) {
-        setBrightnessAt(i, brightness);
-      }
+    void end() {
+      LedModule::end();
     }
 
     //-----------------------------------------------------------------------
@@ -186,6 +165,7 @@ class ScanningModule : public LedModule {
       if (pos >= T_DIGITS) return;
       mBrightnesses[pos] = (brightness >= T_SUBFIELDS)
           ? T_SUBFIELDS : brightness;
+      mIsDigitBrightnessDirty = true;
     }
 
     //-----------------------------------------------------------------------
@@ -239,6 +219,7 @@ class ScanningModule : public LedModule {
      * handler.
      */
     void renderFieldNow() {
+      updateBrightness();
       if (T_SUBFIELDS > 1) {
         displayCurrentFieldModulated();
       } else {
@@ -247,7 +228,8 @@ class ScanningModule : public LedModule {
     }
 
   private:
-    friend class ::ScanningModuleTest_displayCurrentField;
+    friend class ::ScanningModuleTest_isAnyDigitDirty;
+    friend class ::ScanningModuleTest_isBrightnessDirty;
 
     // disable copy-constructor and assignment operator
     ScanningModule(const ScanningModule&) = delete;
@@ -293,6 +275,21 @@ class ScanningModule : public LedModule {
       }
     }
 
+    /**
+     * Transfer the global brightness to the per-digit brightness and update the
+     * appropriate flags.
+     */
+    void updateBrightness() {
+      if (isBrightnessDirty()) {
+        for (uint8_t i = 0; i < T_DIGITS; i++) {
+          setBrightnessAt(i, getBrightness());
+        }
+
+        // Clear the global brightness dirty flag.
+        clearBrightnessDirty();
+      }
+    }
+
   private:
     // The ordering of the fields below partially motivated to save memory on
     // 32-bit processors.
@@ -301,10 +298,10 @@ class ScanningModule : public LedModule {
     const T_LM& mLedMatrix;
 
     /** Pattern for each digit. */
-    volatile uint8_t mPatterns[T_DIGITS];
+    uint8_t mPatterns[T_DIGITS];
 
     /** Brightness for each digit. Unused if T_SUBFIELDS <= 1. */
-    volatile uint8_t mBrightnesses[T_DIGITS];
+    uint8_t mBrightnesses[T_DIGITS];
 
     //-----------------------------------------------------------------------
     // Variables needed by renderFieldWhenReady() to render frames and fields at
@@ -324,6 +321,9 @@ class ScanningModule : public LedModule {
     // Variables needed to keep track of the multiplexing of the digits,
     // and PWM of a single digit.
     //-----------------------------------------------------------------------
+
+    /** Dirty flag for any of the digit-specific brightness. */
+    bool mIsDigitBrightnessDirty;
 
     /**
      * Within the renderFieldNow() method, mCurrentDigit is the current