+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Changelog
 
 * Unreleased
+* 0.8.2 (2021-08-09)
+    * Make `LedModule::getPatternAt()` into a `const` function.
+    * Change `LedModule` methods into non-virtual. There are no more virtual
+      methods in the entire library. Saves 10-60 byte of flash for Module
+      classes on AVR, and another 10-70 bytes of flash for the Writer classes.
+    * Templatize Writer classes on `T_LED_MODULE` instead of hardcoding the
+      dependency to `LedModule`.
+    * Add `isFlushRequired()` to modules which implement the `flush()` method
+      (`Tm1637Module`, `Max7219Module`, `Ht16k33Module`).
+        * Clear dirty flags as needed after `flush()`. Adds about 8 bytes of
+          flash on AVR.
 * 0.8.1 (2021-07-31)
     * Add `HelloTm1636`, `HelloMax7219`, `HelloHt16k33`, and `HelloHc595`
       examples to `examples/` and `README.md`.